# Faster unsafe serialization/deserialization of byte and boolean