The current code does not use unsafe for (de)serialization of bytes and booleans in `UnsafeInput`
and `UnsafeOutput`. Instead, it uses the old mechanism of writing/reading to/from the byte array
buffer for bytes and booleans. Dealing with array poses some overhead compared to unsafe. For
instance, JVM adds boundary checks internally when dealing with arrays. Through detailed
performance profiling it is noticeable that replacing array operations with unsafe operations
for bytes and booleans improves the performance.

For the purpose of performance measurement, we use a micro-benchmark in which a sequence
of bytes/booleans are serialized and written/read to/from a file. The size of the file is twice the
size of the available RAM to reduce/avoid the effect of OS cache. We use UnsafeInput/UnsafeOutput with
a buffer size of 4MB, over FileOutputStream/FileInputStream created from RandomAccessFile
(despite the use of RandomAccessFile, the benchmark only does sequential read/write to file).
Here are the performance results on a machine with Intel Core i7-2.8Ghz and SSD (about 800MB/s read,
and about 700MB/s write bandwidth):

old code using array operation:
read bandwidth: ~650MB/s
write bandwidth: ~550MB/s

changed code using unsafe for byte/boolean:
read bandwidth: ~750 MB/s
write bandwidth: ~650 MB/s

It seems that using unsafe operation improves the IO bandwidth by about 15%.

Test Plan:
mvn clean verify