@romix 2016-06-08
Thanks for submitting this PR! 

Could you also try to add a benchmark? It could be e.g. a JUnit test which always succeeds, but also outputs some performance stats.