@magro 2016-06-09
Just to be sure: this change doesn't affect serialization compatiblity (data written with a previous kryo version can be read with the new one), right?