@romix 2016-06-09
@magro Yes, this is my understanding as well. After all, endianness does not matter if you serialize a single byte, or? ;-)