@heslami 2016-06-10
As for the endianness, the diff does not affect it, single byte/boolean is encoded the same way with different endianness.

As for the benchmark, I'll add a reasonably sized benchmark to show the effect.