@romix 2016-06-22
@heslami Would you have an opportunity to add a benchmark? Then we could merge this PR for the next release.