@heslami 2016-06-23
Sorry for the delay, I was planning to do so in a few days. Currently, I am benchmarking using Kryo in another open source project (Apache Giraph -- code at: https://reviews.facebook.net/D59691). Kryo will be the main (de)serialization mechanism for a key feature in Giraph.